Add moderation filtering to notifications
