+++ src/state/queries/notifications/unread.tsx
@@ -1,13 +1,8 @@
 import React from 'react'
-import {
-  AppBskyNotificationListNotifications,
-  ModerationOpts,
-  moderateProfile,
-  moderatePost,
-} from '@atproto/api'
 import BroadcastChannel from '#/lib/broadcast'
 import {useSession} from '#/state/session'
 import {useModerationOpts} from '../preferences'
+import {shouldFilterNotif} from './util'
 
 const UPDATE_INTERVAL = 30 * 1e3 // 30sec
 
@@ -74,7 +69,7 @@ export function Provider({children}: React.PropsWithChildren<{}>) {
         // count
         const res = await agent.listNotifications({limit: 40})
         const filtered = res.data.notifications.filter(
-          notif => !notif.isRead && !shouldFilter(notif, moderationOpts),
+          notif => !notif.isRead && !shouldFilterNotif(notif, moderationOpts),
         )
         const num =
           filtered.length >= 30
@@ -110,36 +105,4 @@ export function useUnreadNotifications() {
 
 export function useUnreadNotificationsApi() {
   return React.useContext(apiContext)
-}
-
-// TODO this should be in the sdk as moderateNotification -prf
-function shouldFilter(
-  notif: AppBskyNotificationListNotifications.Notification,
-  moderationOpts: ModerationOpts | undefined,
-): boolean {
-  if (!moderationOpts) {
-    return false
-  }
-  const profile = moderateProfile(notif.author, moderationOpts)
-  if (
-    profile.account.filter ||
-    profile.profile.filter ||
-    notif.author.viewer?.muted
-  ) {
-    return true
-  }
-  if (
-    notif.type === 'reply' ||
-    notif.type === 'quote' ||
-    notif.type === 'mention'
-  ) {
-    // NOTE: the notification overlaps the post enough for this to work
-    const post = moderatePost(notif, moderationOpts)
-    if (post.content.filter) {
-      return true
-    }
-  }
-  // TODO: thread muting is not being applied
-  // (this requires fetching the post)
-  return false
 }
